refactor: Move formatUtils to utils folder
